moved sync_webapi function to utils file
